+++ tests/PassportValidatorTest.php
@@ -101,4 +101,10 @@ it('returns the correct birthdate', function($data, $expectation) {
     [['serial' => '3491039407', 'birth' => '9603080', 'gender' => 'F', 'expire' => '4707145', 'nationality' => 'D', 'checksum' => '8'], '1996-03-08'],
     [['serial' => '1087601249', 'birth' => '9604249', 'gender' => 'F', 'expire' => '4612241', 'nationality' => 'D', 'checksum' => '2'], '1996-04-24'],
     [['serial' => '9620118966', 'birth' => '9607170', 'gender' => 'F', 'expire' => '4701110', 'nationality' => 'D', 'checksum' => '6'], '1996-07-17'],
+    [['serial' => '1610772259', 'birth' => '3907100', 'gender' => 'F', 'expire' => '4712233', 'nationality' => 'D', 'checksum' => '6'], '1939-07-10'],
+    [['serial' => '3105880104', 'birth' => '6704148', 'gender' => 'F', 'expire' => '4706230', 'nationality' => 'D', 'checksum' => '4'], '1967-04-14'],
+    [['serial' => '3105880104', 'birth' => '2105312', 'gender' => 'F', 'expire' => '4706230', 'nationality' => 'D', 'checksum' => '2'], '2021-05-31'],
+    [['serial' => '3105880104', 'birth' => '2205315', 'gender' => 'F', 'expire' => '4706230', 'nationality' => 'D', 'checksum' => '2'], '2022-05-31'],
+    [['serial' => '3105880104', 'birth' => '0201221', 'gender' => 'F', 'expire' => '4706230', 'nationality' => 'D', 'checksum' => '8'], '2002-01-22'],
+    [['serial' => '3105880104', 'birth' => '1602294', 'gender' => 'F', 'expire' => '4706230', 'nationality' => 'D', 'checksum' => '6'], '2016-02-29'],
 ]);