+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 1.1.0: 2022-05-31
+* Add dev package requirements
+* Add more test cases
+* Fix bug with incorrect birth years
+* Refactor code to make use of trait
+
 ### 1.0.1: 2020-10-04
 * Change package minimal-stability